betta persistence is good
