+++ internal/tqw/parse.go
@@ -1,339 +1,26 @@
-package game
+package tqw
 
 import (
 	"fmt"
 	"strings"
 
-	"github.com/BurntSushi/toml"
+	"github.com/dekarrin/tunaq/internal/game"
 )
 
-type tqwManifest struct {
-	Format string   `toml:"format"`
-	Type   string   `toml:"type"`
-	Files  []string `toml:"files"`
-}
-
-type twqNPC struct {
-	Label       string          `toml:"label"`
-	Name        string          `toml:"name"`
-	Pronouns    string          `toml:"pronouns"`
-	PronounSet  tqwPronounSet   `toml:"customPronounSet"`
-	Description string          `toml:"description"`
-	Start       string          `toml:"start"`
-	Movement    tqwRoute        `toml:"movement"`
-	Dialog      []tqwDialogStep `toml:"dialog"`
-}
-
-func (tn twqNPC) toNPC() NPC {
-	npc := NPC{
-		Label:       strings.ToUpper(tn.Label),
-		Name:        tn.Name,
-		Pronouns:    tn.PronounSet.toPronounSet(),
-		Description: tn.Description,
-		Start:       tn.Start,
-		Movement:    tn.Movement.toRoute(),
-		Dialog:      make([]DialogStep, len(tn.Dialog)),
-	}
-
-	for i := range tn.Dialog {
-		npc.Dialog[i] = tn.Dialog[i].toDialogStep()
-	}
-
-	return npc
-}
-
-type tqwRoute struct {
-	Action         string   `toml:"action"`
-	Path           []string `toml:"path"`
-	ForbiddenRooms []string `toml:"forbiddenRooms"`
-	AllowedRooms   []string `toml:"allowedRooms"`
-}
-
-func (tr tqwRoute) toRoute() Route {
-	act, ok := RouteActionsByString[strings.ToUpper(tr.Action)]
-	if !ok {
-		act = RouteStatic
-	}
-
-	r := Route{
-		Action:         act,
-		Path:           make([]string, len(tr.Path)),
-		ForbiddenRooms: make([]string, len(tr.ForbiddenRooms)),
-		AllowedRooms:   make([]string, len(tr.AllowedRooms)),
-	}
-
-	copy(r.Path, tr.Path)
-	copy(r.ForbiddenRooms, tr.ForbiddenRooms)
-	copy(r.AllowedRooms, tr.AllowedRooms)
-
-	return r
-}
-
-type tqwDialogStep struct {
-	Action   string     `toml:"action"`
-	Label    string     `toml:"label"`
-	Content  string     `toml:"content"`
-	Response string     `toml:"response"`
-	Choices  [][]string `toml:"choices"`
-}
-
-func (tds tqwDialogStep) toDialogStep() DialogStep {
-	act, ok := DialogActionsByString[strings.ToUpper(tds.Action)]
-	if !ok {
-		act = DialogLine
-	}
-
-	ds := DialogStep{
-		Action:   act,
-		Label:    tds.Label,
-		Content:  tds.Content,
-		Response: tds.Response,
-		Choices:  make(map[string]string),
-	}
-
-	for _, ch := range tds.Choices {
-		if len(ch) < 2 {
-			continue
-		}
-
-		choice := ch[0]
-		dest := ch[1]
-		ds.Choices[choice] = dest
-	}
-
-	return ds
-}
-
-type tqwPronounSet struct {
-	Nominative string `toml:"nominative"`
-	Objective  string `toml:"objective"`
-	Possessive string `toml:"possessive"`
-	Determiner string `toml:"determiner"`
-	Reflexive  string `toml:"reflexive"`
-}
-
-func pronounSetToTWQ(ps PronounSet) tqwPronounSet {
-	tp := tqwPronounSet{
-		Nominative: ps.Nominative,
-		Objective:  ps.Objective,
-		Possessive: ps.Possessive,
-		Determiner: ps.Determiner,
-		Reflexive:  ps.Reflexive,
-	}
-
-	return tp
-}
-
-func (tp tqwPronounSet) toPronounSet() PronounSet {
-	ps := PronounSet{
-		Nominative: strings.ToUpper(tp.Nominative),
-		Objective:  strings.ToUpper(tp.Objective),
-		Possessive: strings.ToUpper(tp.Possessive),
-		Determiner: strings.ToUpper(tp.Determiner),
-		Reflexive:  strings.ToUpper(tp.Reflexive),
-	}
-
-	// default to they/them fills
-	if ps.Nominative == "" {
-		ps.Nominative = "THEY"
-	}
-	if ps.Objective == "" {
-		ps.Objective = "THEM"
-	}
-	if ps.Possessive == "" {
-		ps.Possessive = "THEIRS"
-	}
-	if ps.Determiner == "" {
-		ps.Determiner = "THEIR"
-	}
-	if ps.Reflexive == "" {
-		ps.Reflexive = "THEMSELF"
-	}
-
-	return ps
-}
-
-type tqwItem struct {
-	Label       string   `toml:"label"`
-	Name        string   `toml:"name"`
-	Description string   `toml:"description"`
-	Aliases     []string `toml:"aliases"`
-}
-
-func (ti tqwItem) toItem() Item {
-	item := Item{
-		Label:       ti.Label,
-		Name:        ti.Name,
-		Description: ti.Description,
-		Aliases:     make([]string, len(ti.Aliases)),
-	}
-
-	copy(item.Aliases, ti.Aliases)
-
-	return item
-}
-
-type tqwEgress struct {
-	DestLabel     string   `toml:"destLabel"`
-	Description   string   `toml:"description"`
-	TravelMessage string   `toml:"travelMessage"`
-	Aliases       []string `toml:"aliases"`
-}
-
-func (te tqwEgress) toEgress() Egress {
-	eg := Egress{
-		DestLabel:     te.DestLabel,
-		Description:   te.Description,
-		TravelMessage: te.TravelMessage,
-		Aliases:       make([]string, len(te.Aliases)),
-	}
-
-	copy(eg.Aliases, te.Aliases)
-
-	return eg
-}
-
-type tqwRoom struct {
-	Label       string      `toml:"label"`
-	Name        string      `toml:"name"`
-	Description string      `toml:"description"`
-	Exits       []tqwEgress `toml:"exits"`
-	Items       []tqwItem   `toml:"items"`
-}
-
-func (tr tqwRoom) toRoom() Room {
-	r := Room{
-		Label:       tr.Label,
-		Name:        tr.Name,
-		Description: tr.Description,
-		Exits:       make([]Egress, len(tr.Exits)),
-		Items:       make([]Item, len(tr.Items)),
-		NPCs:        make(map[string]*NPC),
-	}
-
-	for i := range tr.Exits {
-		r.Exits[i] = tr.Exits[i].toEgress()
-	}
-	for i := range tr.Items {
-		r.Items[i] = tr.Items[i].toItem()
-	}
-
-	return r
-}
-
-type tqwFileInfo struct {
-	Format string `toml:"format"`
-	Type   string `toml:"type"`
-}
-
-type tqwWorld struct {
-	Start string `toml:"start"`
-}
-
-// tqwWorldData is the top-level structure containing all keys in a complete TQW
-// 'DATA' type file.
-type tqwWorldData struct {
-	Format   string                   `toml:"format"`
-	Type     string                   `toml:"type"`
-	Rooms    []tqwRoom                `toml:"rooms"`
-	World    tqwWorld                 `toml:"world"`
-	NPCs     []twqNPC                 `toml:"npcs"`
-	Pronouns map[string]tqwPronounSet `toml:"pronouns"`
-}
-
-// ParseManifestFromTOML takes in raw TOML bytes, reads it for manifest data,
-// and returns the files in the manifest.
-func ParseManifestFromTOML(manifestData []byte) (Manifest, error) {
-	var loaded tqwManifest
-
-	if unmarshalErr := toml.Unmarshal(manifestData, &loaded); unmarshalErr != nil {
-		return Manifest{}, unmarshalErr
-	}
-
-	if strings.ToUpper(loaded.Format) != "TUNA" {
-		return Manifest{}, fmt.Errorf("in header: 'format' key must exist and be set to 'TUNA'")
-	}
-
-	if strings.ToUpper(loaded.Type) != "MANIFEST" {
-		return Manifest{}, fmt.Errorf("in header: type set to %q, not \"MANIFEST\"", loaded.Type)
-	}
-
+func parseManifest(tqw tqwManifest) (Manifest, error) {
 	manif := Manifest{
-		Files: loaded.Files,
+		Files: tqw.Files,
 	}
 
 	return manif, nil
 }
 
-// ParseWorldFromTOML takes in raw TOML bytes, reads it for a world definition,
-// and returns the rooms as well as the label of the starting room.
-//
-// Note: Uses module-global variables as part of operation. Absolutely not
-// thread-safe and calling more than once concurrently will lead to unexpected
-// results.
-func ParseWorldDataFromTOML(tomlData []byte) (WorldData, error) {
-	translatedData, err := UnmarshalTOMLWorldData(tomlData)
-	if err != nil {
-		return WorldData{}, err
-	}
-	return parseUnmarshaledData(translatedData)
-}
-
-// UnmarshalTOMLWorldData unmarshals but does not parse or check the loaded
-// world data.
-func UnmarshalTOMLWorldData(tomlData []byte) (tqwWorldData, error) {
-	var tqw tqwWorldData
-	if tomlErr := toml.Unmarshal(tomlData, &tqw); tomlErr != nil {
-		return tqw, tomlErr
-	}
-
-	if strings.ToUpper(tqw.Format) != "TUNA" {
-		return tqw, fmt.Errorf("in header: 'format' key must exist and be set to 'TUNA'")
-	}
-	if strings.ToUpper(tqw.Type) != "DATA" {
-		return tqw, fmt.Errorf("in header: 'type' must exist and be set to 'DATA'")
-	}
-
-	return tqw, nil
-}
-
-// scan the first lines for format info before doing anything else
-func scanFileInfo(data []byte) (tqwFileInfo, error) {
-	// only run the toml parser up to the end of the top-lev table
-
-	var topLevelEnd int = -1
-	var onNewLine bool
-	for b := range data {
-		if onNewLine {
-			if data[b] == '[' {
-				topLevelEnd = b
-				break
-			}
-		}
-
-		if data[b] == '\n' {
-			onNewLine = true
-		} else {
-			onNewLine = false
-		}
-	}
-
-	scanData := data
-	if topLevelEnd != -1 {
-		scanData = data[:topLevelEnd]
-	}
-
-	var info tqwFileInfo
-	err := toml.Unmarshal(scanData, &info)
-	return info, err
-}
-
-func parseUnmarshaledData(tqw tqwWorldData) (WorldData, error) {
+func parseWorldData(tqw tqwWorldData) (WorldData, error) {
 	var err error
 
 	var world WorldData
 	world.Start = tqw.World.Start
-	world.Rooms = make(map[string]*Room)
+	world.Rooms = make(map[string]*game.Room)
 
 	for idx, r := range tqw.Rooms {
 		if roomErr := validateRoomDef(r); roomErr != nil {
@@ -349,10 +36,10 @@ func parseUnmarshaledData(tqw tqwWorldData) (WorldData, error) {
 	}
 
 	pronouns := map[string]tqwPronounSet{
-		"SHE/HER":   pronounSetToTWQ(PronounsFeminine),
-		"HE/HIM":    pronounSetToTWQ(PronounsMasculine),
-		"THEY/THEM": pronounSetToTWQ(PronounsNonBinary),
-		"IT/ITS":    pronounSetToTWQ(PronounsItIts),
+		"SHE/HER":   pronounSetToTWQ(game.PronounsFeminine),
+		"HE/HIM":    pronounSetToTWQ(game.PronounsMasculine),
+		"THEY/THEM": pronounSetToTWQ(game.PronounsNonBinary),
+		"IT/ITS":    pronounSetToTWQ(game.PronounsItIts),
 	}
 
 	// check loaded pronouns
@@ -368,7 +55,7 @@ func parseUnmarshaledData(tqw tqwWorldData) (WorldData, error) {
 		pronouns[name] = ps
 	}
 
-	npcs := make([]NPC, 0)
+	npcs := make([]game.NPC, 0)
 	// parse individual npcs
 	for idx, npc := range tqw.NPCs {
 		if npc.Movement.Action == "" {
@@ -418,7 +105,7 @@ func parseUnmarshaledData(tqw tqwWorldData) (WorldData, error) {
 	// ensure that all npc routes are valid, that convo trees make sense, then place NPCs in their start rooms
 	// also ensure that all labels are unique among NPCs.
 	seenNPCLabels := map[string]int{}
-	pf := Pathfinder{World: world.Rooms}
+	pf := game.Pathfinder{World: world.Rooms}
 	for idx, npc := range npcs {
 		// check labels
 		if seenInIndex, ok := seenNPCLabels[npc.Label]; ok {
@@ -434,7 +121,7 @@ func parseUnmarshaledData(tqw tqwWorldData) (WorldData, error) {
 
 		// then check route based on movement type
 		switch npc.Movement.Action {
-		case RoutePatrol:
+		case game.RoutePatrol:
 			// can npc get to initial position?
 			err = pf.ValidatePath(append([]string{npc.Start}, npc.Movement.Path[0]), false)
 			if err != nil {
@@ -446,7 +133,7 @@ func parseUnmarshaledData(tqw tqwWorldData) (WorldData, error) {
 			if err != nil {
 				return world, fmt.Errorf("validating: npcs[%d (%q)]: %w", idx, npc.Label, err)
 			}
-		case RouteWander:
+		case game.RouteWander:
 			for roomIdx, roomLabel := range npc.Movement.AllowedRooms {
 				_, ok := world.Rooms[roomLabel]
 				if !ok {
@@ -491,7 +178,7 @@ func parseUnmarshaledData(tqw tqwWorldData) (WorldData, error) {
 					}
 				}
 			}
-		case RouteStatic:
+		case game.RouteStatic:
 			// nothing more to validate, they don't move
 		default:
 			return world, fmt.Errorf("validating: npcs[%d (%q)]: movement: action: unknown action type '%v'", idx, npc.Label, npc.Movement.Action)
@@ -515,7 +202,7 @@ func parseUnmarshaledData(tqw tqwWorldData) (WorldData, error) {
 
 		// check convo tree for choice label validity
 		for diaIdx, diaStep := range npc.Dialog {
-			if diaStep.Action == DialogChoice {
+			if diaStep.Action == game.DialogChoice {
 				choiceNum := -1
 				for _, dest := range diaStep.Choices {
 					choiceNum++
@@ -568,21 +255,21 @@ func validateNPCDef(npc twqNPC, topLevelPronouns map[string]tqwPronounSet) error
 
 func validateDialogStepDef(ds tqwDialogStep) error {
 	diaUpper := strings.ToUpper(ds.Action)
-	dia, ok := DialogActionsByString[diaUpper]
+	dia, ok := game.DialogActionsByString[diaUpper]
 
 	if !ok {
 		return fmt.Errorf("action: must be one of 'LINE', 'CHOICE', or 'END', not %q", diaUpper)
 	}
 
 	switch dia {
-	case DialogLine:
+	case game.DialogLine:
 		if len(ds.Choices) > 0 {
 			return fmt.Errorf("'LINE' dialog step type does not use 'choices' key")
 		}
 		if ds.Content == "" {
 			return fmt.Errorf("'LINE' dialog step type requires a string as value of 'content' property")
 		}
-	case DialogChoice:
+	case game.DialogChoice:
 		if len(ds.Choices) < 2 {
 			return fmt.Errorf("'CHOICE' dialog step type must have a list with at least 2 choices as value of 'choices' property")
 		}
@@ -592,7 +279,7 @@ func validateDialogStepDef(ds tqwDialogStep) error {
 		if ds.Content == "" {
 			return fmt.Errorf("'CHOICE' dialog step type requires a string as value of 'content' property")
 		}
-	case DialogEnd:
+	case game.DialogEnd:
 		if ds.Response != "" {
 			return fmt.Errorf("'END' dialog step type does not use 'response' property")
 		}
@@ -612,14 +299,14 @@ func validateDialogStepDef(ds tqwDialogStep) error {
 
 func validateRouteDef(ps tqwRoute) error {
 	actUpper := strings.ToUpper(ps.Action)
-	act, ok := RouteActionsByString[actUpper]
+	act, ok := game.RouteActionsByString[actUpper]
 
 	if !ok {
 		return fmt.Errorf("action: must be one of 'STATIC', 'PATROL', or 'WANDER', not %q", actUpper)
 	}
 
 	switch act {
-	case RoutePatrol:
+	case game.RoutePatrol:
 		if len(ps.Path) < 2 {
 			return fmt.Errorf("'PATROL' route type must have a list with at least 2 rooms as value of 'path' property")
 		}
@@ -629,11 +316,11 @@ func validateRouteDef(ps tqwRoute) error {
 		if len(ps.ForbiddenRooms) > 0 {
 			return fmt.Errorf("'PATROL' route type does not use 'forbiddenRooms' property")
 		}
-	case RouteWander:
+	case game.RouteWander:
 		if len(ps.Path) > 0 {
 			return fmt.Errorf("'WANDER' route type does not use 'path' property")
 		}
-	case RouteStatic:
+	case game.RouteStatic:
 		if len(ps.Path) > 0 {
 			return fmt.Errorf("'STATIC' route type does not use 'path' property")
 		}
